getUserName 구조 이동
